refactor(local-ci): extract job_queue.py (R2-1 seventh cut)
Move 3 queue persistence helpers (normalize_job, load_queue_unlocked,
save_queue_unlocked) out of local_ci.py into a sibling job_queue.py
module. Renamed from queue.py to avoid collision with the stdlib
queue module.

- local_ci.py: 10,870 → 10,843 lines (-27 net).
- job_queue.py: 68 lines.
- Depends on state_paths (queue_path), io_utils (atomic_write_text),
  normalize (normalize_priority, normalize_validation_mode),
  provenance (normalize_provenance) — leverages the seams the previous
  cuts established.
- The lock-acquiring `load_queue` (which also reconciles running jobs)
  stays in local_ci.py because it pulls in the running-job state
  machine — not part of this seam.

Tests: 213/213 local_ci tests pass.

Running R2-1 total: 7 modules extracted (state_paths, normalize,
git_helpers, io_utils, footprint, provenance, job_queue).
local_ci.py shrunk 11,285 → 10,843 lines (-442, -3.9%).
